+++ tests/Suggest/PhraseTest.php
@@ -76,37 +76,6 @@ class PhraseTest extends BaseTest
         $this->assertEquals('elasticsearch is bonsai cool', $suggests['suggest1'][0]['options'][0]['text']);
     }
 
-    /**
-     * @group functional
-     * @group legacy
-     */
-    public function testPhraseSuggestWithBackwardsCompatibility(): void
-    {
-        $this->expectDeprecation('Since ruflin/elastica 7.2.0: The "Elastica\Suggest\Phrase::addCandidateGenerator()" method is deprecated, use the "addDirectGenerator()" method instead. It will be removed in 8.0.');
-
-        $phraseSuggest = (new Phrase('suggest1', 'text'))
-            ->setText('elasticsearch is bansai coor')
-            ->setAnalyzer('simple')
-            ->setHighlight('<suggest>', '</suggest>')
-            ->setStupidBackoffSmoothing(Phrase::DEFAULT_STUPID_BACKOFF_DISCOUNT)
-            ->addCandidateGenerator(new DirectGenerator('text'))
-        ;
-
-        $suggest = (new Suggest())
-            ->addSuggestion($phraseSuggest)
-        ;
-
-        $index = $this->_getIndexForTest();
-        $result = $index->search($suggest);
-        $suggests = $result->getSuggests();
-
-        // 3 suggestions should be returned: One in which both misspellings are corrected, and two in which only one misspelling is corrected.
-        $this->assertCount(3, $suggests['suggest1'][0]['options']);
-
-        $this->assertEquals('elasticsearch is <suggest>bonsai cool</suggest>', $suggests['suggest1'][0]['options'][0]['highlighted']);
-        $this->assertEquals('elasticsearch is bonsai cool', $suggests['suggest1'][0]['options'][0]['text']);
-    }
-
     protected function _getIndexForTest(): Index
     {
         $index = $this->_createIndex();